RR-1238 - Refactor date calculation tests for Review Schedules
